+++ pkg/utils/utils.go
@@ -342,3 +342,29 @@ func GetLocalIPAddresses() []string {
 
 	return addresses
 }
+
+// GetShipyardIPAndHostname returns the IP Address of the machine
+// running shipyard
+func GetShipyardIPAndHostname() (string, string) {
+	addrs, err := net.InterfaceAddrs()
+
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	var currentIP, currentNetworkHardwareName string
+
+	for _, address := range addrs {
+
+		// check the address type and if it is not a loopback the display it
+		// = GET LOCAL IP ADDRESS
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				fmt.Println("Current IP address : ", ipnet.IP.String())
+				currentIP = ipnet.IP.String()
+			}
+		}
+	}
+
+	return currentIP, currentNetworkHardwareName
+}